# optimize: Skip unnecessary costly table content building