- `method_table_content` is built only for `/api/*`
- `actions_table_content` is built only for `/iam/*`
- `api_count` is always counted since it is shown both in `/api/*` and `/iam/*`

page | before | after
---|---|---
/iam/s3 |  40 ms| 22 ms
/api/s3 | 38 ms | 21 ms